Implement support for site advertisements
http://www.bloc.io/users/danya-smith/checkpoints/1639?roadmap_section_id=120

It is time to monetize the Bloccit site using advertisements. In order to provide support for ads, we need to collect the title, ad copy and price for each advertisement. The Advertisement model will need supporting views for the create, index, new and show actions. These features were implemented using TDD.

* Created the Advertisement model
* Migrated the Advertisments table to the database
* Implemented the controller
* Implemented the create, index, new and show views