Allow users to update and delete posts
https://www.bloc.io/users/danya-smith/checkpoints/1640?roadmap_section=120

Using TDD, I wrote test cases for edit and update actions before modifying the Posts controller to include edit and update action methods. After implementing the ability to edit and update posts, I modified the edit and show views so that the user is presented with a link to the edit form from the show view.

Next, using TDD I wrote test cases for the destroy action before modifying the Posts controller to include a destroy action method. After implementing the ability to delete a post, I modified the show view so that the user is presented with a button to use to delete the selected post.

On branch feature/Posts_Update_and_Delete
Changes to be committed:

* modified:   app/controllers/posts_controller.rb
* modified:   app/models/post.rb
* modified:   app/views/posts/edit.html.erb
* modified:   app/views/posts/show.html.erb
* modified:   spec/controllers/posts_controller_spec.rb