Adds moderator role and implements authorization rules
https://www.bloc.io/users/danya-smith/checkpoints/1644?roadmap_section=120

In this checkpoint, we learn that we can add a Role field to the database
and use the enum structure to populate it with roles. In this assignment,
we add the :moderator role to existing roles, and provide it with the
ability to edit Topics, and edit and create Posts.