Partial Implementation of Checkpoint 46
https://www.bloc.io/users/danya-smith/checkpoints/1648?roadmap_section_id=120

I started the checkpoint to implement Labels for topics and posts. The use of the simple_form gem in prior checkpoints has caused my solution to deviate from the roadmap. I was able to get the forms to display properly, but there seems to be something missing from my partials implementation. While the forms display properly, the Edit form does not display the existing content in the form elements. Need to either rollback the use of simple form and complete this using the recommended method, or get this reviewed.