Allows Posts and Topics to be Tagged with Labels
https://www.bloc.io/users/danya-smith/checkpoints/1648?roadmap_section_id=120

In this checkpoint, we manually create two models: label and labeling. These
models create a polymorphic association: an interface that any other model can
use. Labels have a has_many relationship to both Topics and Posts. We display
labels on both Topic and Post show views, and create a Label view which lists
all of the Topics and Posts associated with a given label.

Two things caused issues for me with this assignment:

I tried to create a before_action in the TopicsController that assigns @topic to
equal Topic.find(params[:topic_id]). This broke a number of tests, because
before_actions can accept variables, but not assignment statements.

Using the simple_form gem was not a part of this project. As such, the code
samples do not explain the aspects of the simple_form gem that are required.
I needed to wrap the value property in brackets with the input_html: symbol.

I tried to hard-code the paths I needed from my raked routes to resolve an
issue I was having with labels not displaying properly (this was actually solved
by using the input_html: symbol, as documented in the simple_form gem). Instead,
I used the model names (@post or @topic) but had to make sure they were passed
in as local variables. In one case, I had to pass the value of @topic through
the edit action in the PostsController.